add upload, delete and get file

+++ src/modules/folder/folder.service.ts
@@ -1,35 +1,46 @@
-import { EntityManager, EntityRepository, wrap } from '@mikro-orm/postgresql';
+import { EntityManager, EntityRepository, RequiredEntityData, wrap } from '@mikro-orm/postgresql';
 import { Folder } from "./folder.entity";
 export class FolderService  {
-     constructor(
+    constructor(
         private readonly em:EntityManager,
-        private readonly folderRepo:EntityRepository<Folder>
-
-     ){
+        private readonly folderRepo:EntityRepository<Folder>,
+    ){
         this.em = em,
         this.folderRepo = folderRepo
 
-     }
-    createFolder(folder){
+    }
 
+    private dontTochError = new Error('don`t tach this folder')
+    private notExistError = new Error('folder not exist')
+    
+    createFolder(folder: RequiredEntityData<Folder, never, false>){
         this.em.create(Folder,folder)
         return this.em.flush()
     }
-    async GetFolder(id:number|null){
+    async GetFolder(id:string|null){
         const res = await this.folderRepo.createQueryBuilder('f')
         .where({'f.id':id})
         .leftJoinAndSelect('f.files','file',{})
         .leftJoinAndSelect('f.folders','folder',{})
         .getResult()
+        if (res.length == 0){
+            throw this.notExistError
+        }
         return res
     }
-    async DeleteFolder(id:number){
-        const folder = await this.em.find(Folder, {id});
+    async DeleteFolder(id){
+        const folder = await this.em.findOneOrFail(Folder, {id});
+        if(folder.name == 'Root'){
+            throw this.dontTochError
+        }
         this.em.remove(folder);
         await this.em.flush();
     }
-    async UpdateFolder(id:number,name:string){
-        const folder = await this.folderRepo.findOne({id});
+    async UpdateFolder(id,name){
+        const folder = await this.folderRepo.findOneOrFail({id});
+        if(folder.name == 'Root'){
+            throw this.dontTochError
+        }
         if (folder){
             wrap(folder).assign({name})
             await this.em.flush()
@@ -38,7 +49,6 @@ export class FolderService  {
         const folders = await this.em.createQueryBuilder(Folder,'folder')
         .leftJoinAndSelect('folder.folders','f')
         .getResult()
-   
         const root = folders.shift()
         root.folders = this.createTree(folders,root.id)
 
@@ -60,4 +70,3 @@ export class FolderService  {
         return newArr
     }
 }
-
